DetectChief.ModeloDeDeteccion/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
transformers==4.40.1
torch==2.2.2
pika==1.3.2
Pillow==10.3.0